[YS-462] refactor: 학교 메일 인증 입력 컴포넌트 분리
